Game state implementation.
Game state is working, made a different build bat to build and test it
as I have no game state to actually play the game. Just edit luckyB.cpp
to change which game version, and run the normal bat file for the
original version. Now need to get the main menu layout done with buttons
and everything, and get an actual game state to start playing the game.