Moved handle input from player to game.
Moved the handle input stuff from the player class into the game state.
We were passing almost all the variables from the game into the handle
input class in player, so why not just implement it in the game class. I
also made a list of players. I am still working on better organization
on the stuff i moved over.